Move coast function to coast.py
